DENISE modularization part 5
