+++ src/globvar.h
@@ -47,12 +47,9 @@ float SRTRADIUS, EXP_TAPER_GRAD_HOR;
 
 int SPATFILTER, SPAT_FILT_SIZE, SPAT_FILT_1, SPAT_FILT_ITER;
 int INV_RHO_ITER, INV_VS_ITER, INV_VP_ITER;
-int MIN_ITER, GRAD_FORM;
+int MIN_ITER, GRAD_FORM, IDXI, IDYI, NTDTINV, NXNYI;
 
 char INV_MODELFILE[STRING_SIZE];
-int nfstart, nf;
-
-int nfstart_jac, nf_jac;
 
 float VPUPPERLIM, VPLOWERLIM, VSUPPERLIM, VSLOWERLIM, RHOUPPERLIM, RHOLOWERLIM;
 
@@ -65,7 +62,7 @@ char PARA[STRING_SIZE];
 int TIME_FILT, ORDER, SHOTINC;
 float FC_START, FC_END, FC_INCR;
 
-int LNORM, DTINV, NXNYI, NTDTINV, IDXI, IDYI;
+int LNORM, DTINV;
 
 int STEPMAX;
 float EPS_SCALE, SCALEFAC;
@@ -125,3 +122,5 @@ int ENV;
 /* parameters for towed streamer */
 int N_STREAMER;
 float REC_INCR_X, REC_INCR_Y;
+
+int PHYSICS;
